Extend representation of wui specification
Reuse the old `WuiSpec a` defintion for the new representation `type WuiLensSpec a = Maybe a -> WuiSpec a`. An optional intial value can be passed for each specification instead of the explicit value that is passed by the generation of the form. With this optional value, `transformWSpec` can be defined with lenses: `transformWSpec :: Lens a b -> WuiLensSpec a -> WuiLensSpec b`.
Furthermore, the examples in `FormLenses.curry`  are adopted to fit the new WUI interface.