Add new implementation of lenses
This implementation allows to only define put-based lenses and the corresponding `get` function is defined in terms of `put`. The defintion of `get` is generated by the builtin-search. Nevertheless, not every put-based lens definition has only one corresponding `get`, therefore, we use a helper function to check, if only one result is generated.