Add implementation of test framework `EasyCheck`
The framework was presented in the paper "EasyCheck - Test data for free" (http://www-ps.informatik.uni-kiel.de/~sebf/data/pub/flops08.pdf); this implementation is an adapted version of the implementation found in the old svn-repo, i.e. the svn-repo that was imported to git-ps. The additional module `SearchTreeTraversal` is used by the `EasyCheck` implementation and was also found in the old svn-repo and adapted for further usage.
Last but not least, `Test.curry` contains the examples from the paper.