Add new test functions and adjust tests according to domains.
The new test functions correspond to `PutDet` and `PutStab`, two laws that are formulated on the `put` definition of a lens only. Additionally, adjust the original test functions to only consider valid domains, i.e. if a `put` or `get` functions is not defined for the generated values, we do not check the law, but generate new values instead.