Add module "RegExp.curry" for regular expression matching in strings based on lenses
Three kinds of lens defintions for regular expressions are definied:

`type LensReg   = RegExp -> Lens String String`

This lens definition replaces a substring in a given string for all matched index positions according to the given regular expression. The missing list type is compensated in sake of a nondeterministic result. The `get` direction is also nondeterministic and yields all matches according to the regular expression for a given string.

`type LensReg1  = RegExp -> Lens String (Int,String)`

This version replaces only valid matches for the given index position, i.e. fails if the index position does not match the regular expression.

`type LensReg1M = RegExp -> Lens (Maybe String) (Int,String)`
The third type of lens for regular expressions is neither nondeterministicly nor partially defined, it yields `Nothing` if the given index position and string do not match according to the regular expression.

In this first version, the type for regular expression is defined as
`type RegExp = String -> (Bool,String)`,
where `Bool` indicates if the string matches successfully and the second component is the matched substring.