Add new implementation for replacing/pretty-printing
The new implementation uses the underlying type

`type Reparse a = String -> (a,String) -> (String,String,String)`

that tracks the result of the replacement (1), the remaining string (2) and a string that is concatenated at the end (3). This implementation follows a more functional approach in comparision with the previous implementation (Replace.curry), which followed a logical approach.